[front/lib/email] - feature: implement email sending with configurable templates
 - Added a new function to support sending emails using templates from configuration
 - Moved `sendEmailWithTemplate` function to the centralized email module for reusability

[front/pages/api/w/[wId]/data_sources/request_access] - refactor: update request access to use centralized email sending function

 - Replaced the local `sendEmailWithTemplate` implementation with the one from the `email` module
 - Cleaned up the imports and dependencies related to email sending to use the new function
